chore(team): split mailbox store paths
