+++ src/team/manager.rs
@@ -8,6 +8,7 @@ mod mailbox;
 mod mailbox_payloads;
 mod mailbox_reply_obligations;
 mod mailbox_service;
+mod mailbox_store;
 mod memory_flush;
 mod message_archive;
 mod payload_utils;
